+++ pydantic_resolve/resolver.py
@@ -14,20 +14,15 @@ from aiodataloader import DataLoader
 
 def LoaderDepend(  # noqa: N802
     dependency: Optional[Callable[..., Any]] = None,
-    transform: Optional[Callable[..., Any]] = None 
 ) -> Any:
-    return Depends(
-        dependency=dependency,
-        transform=transform)
+    return Depends(dependency=dependency)
 
 class Depends:
     def __init__(
         self, 
         dependency: Optional[Callable[..., Any]] = None,
-        transform: Optional[Callable[..., Any]] = None
     ):
         self.dependency = dependency
-        self.transform = transform
 
 T = TypeVar("T")
 
@@ -50,13 +45,9 @@ class Resolver:
                 # Base: DataLoader or batch_load_fn
                 # transform: data transform function
                 Base = v.default.dependency
-                transform = v.default.transform
 
                 # module.kls to avoid same kls name from different module
                 cache_key = f'{v.default.dependency.__module__}.{v.default.dependency.__name__}'
-                if transform:
-                    # lambda's name is "lambda", not distinctive.
-                    cache_key = f'{cache_key}:{transform.__module__}.{id(transform)}'  
 
                 cache_provider = self.ctx.get()
                 hit = cache_provider.get(cache_key, None)
@@ -67,18 +58,7 @@ class Resolver:
                     # create loader instance 
                     if isclass(Base):
                         # if extra transform provides
-                        if transform:
-
-                            # replace with new batch_fn with transform function
-                            async def batch_load_fn(self, *args):
-                                items = await Base.batch_load_fn(self, *args)  # type:ignore
-                                items = [transform(item) for item in items]  # type:ignore
-                                return items
-
-                            NewBase = replace_method(Base, 'DataLoader', 'batch_load_fn', batch_load_fn)
-                            loader = NewBase()
-                        else:
-                            loader = Base()
+                        loader = Base()
 
                         # and pick config from 'loader_filters' param, only for DataClass
                         filter_config_provider = self.loader_filters_ctx.get()
@@ -98,11 +78,7 @@ class Resolver:
 
                     # build loader from batch_load_fn, filters config is impossible
                     else:
-                        if transform:
-                            _batch_load_fn = transformer_decorator(transform)(Base)  # type:ignore
-                        else:
-                            _batch_load_fn = Base
-                        loader = DataLoader(batch_load_fn=_batch_load_fn) # type:ignore
+                        loader = DataLoader(batch_load_fn=Base) # type:ignore
 
                     cache_provider[cache_key] = loader
                     self.ctx.set(cache_provider)